requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --annotation-style=line --output-file=requirements.txt requirements.in
#

certifi==2022.6.15        # via requests
charset-normalizer==2.1.0  # via requests
commonmark==0.9.1         # via rich
idna==3.3                 # via requests
pygments==2.12.0          # via rich
requests==2.28.1          # via -r requirements.in
rich==12.5.1              # via -r requirements.in
ruamel-yaml==0.17.21      # via -r requirements.in
ruamel-yaml-clib==0.2.6   # via ruamel-yaml
urllib3==1.26.10          # via requests
